+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Clarion Extension are documented here.
 
 ## Recent Versions
 
+### [1.0.4] - Unreleased
+
+---
+
 ### [1.0.3] - 2026-09-13
 
 **Bug Fixes**